app/requirements.txt
pipdeptree==2.0.0
psycopg2-binary>=2.8
tabulate